Blogcreator update:
Made it possible to save blogs in a special `.blog` format.

The actual data is saved in a HTML file on a location on the server that is not visible for the user.

The blog appears in the user's explorer as a `.blog` file. It is just symbolic.

Right now the user can't edit a blog that is saved, but I will add that functionality later.