Blog bug fix & explorer loading screen:
Fixed a bug where there was no confirmation message when the user tried to save a blogfile with the same name as an existing file.

Added a loading screen to the explorer so when connection is slow the user wont keep clicking on folders and try to access direcrories that don't exist.